requirements.txt
python-dotenv
streamlit
streamlit-extras
altair
PyPDF2
openai
langchain
faiss-cpu
huggingface_hub
InstructorEmbedding
sentence_transformers

